moved loading stuff to App.js
